+++ src/main/java/com/zoopick/server/dto/match/CreateMatchEvent.java
@@ -1,5 +1,3 @@
 package com.zoopick.server.dto.match;
 
-import com.zoopick.server.entity.Item;
-
-public record CreateMatchEvent(Long matchId, Item lostItem, Item foundItem) {}
+public record CreateMatchEvent(Long matchId, Long lostItemId, Long foundItemId) {}
